[Ops] ReConstruct activation ops (#5152)
