+++ lite/tests/kernels/activation_grad_compute_test.cc
@@ -16,6 +16,7 @@
 #include <gtest/gtest.h>
 #include "lite/core/op_registry.h"
 #include "lite/kernels/arm/activation_compute.h"
+#include "lite/kernels/arm/activation_extra_compute.h"
 
 namespace paddle {
 namespace lite {
